crear account

+++ database/migrations/2025_03_04_123510_create_accounts_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('accounts', function (Blueprint $table) {
             $table->id();
             $table->float('balance', 10, 2);
-            $table->integer('number');
+            $table->bigInteger('number');
             $table->string('cvc');
             $table->string('placeholder');
             $table->date('due_date');
